refactor(admin): extract court operations
